[IS-1136] rename testcase/unittest directory
 - fixed to `tests/unit` because of conflict with python `unittest` module
