+++ tests/unit/test_rest_client.py
