+++ tests/unit/test_utils_util.py
@@ -19,7 +19,7 @@
 import logging
 import unittest
 import loopchain.utils as util
-import testcase.unittest.test_util as test_util
+import tests.unit.test_util as test_util
 from loopchain.utils import loggers
 
 loggers.set_preset_type(loggers.PresetType.develop)
